Use the ResponseWriter so the proxy pipeline with start tls stays consistent.
