+++ src/FreeDSx/Ldap/Server/ConnectionHandlerBuilder.php
@@ -33,6 +33,7 @@ use FreeDSx\Ldap\Protocol\Factory\ResponseFactory;
 use FreeDSx\Ldap\Protocol\Factory\ServerProtocolHandlerFactory;
 use FreeDSx\Ldap\Protocol\Queue\Response\MetricsResponseInterceptor;
 use FreeDSx\Ldap\Protocol\Queue\Response\PasswordPolicyResponseInterceptor;
+use FreeDSx\Ldap\Protocol\Queue\Response\ResponseWriter;
 use FreeDSx\Ldap\Protocol\Queue\ServerQueue;
 use FreeDSx\Ldap\Protocol\ServerAuthorization;
 use FreeDSx\Ldap\Protocol\ServerProtocolHandler;
@@ -412,7 +413,7 @@ final class ConnectionHandlerBuilder implements ConnectionHandlerBuilderInterfac
             ],
             new HandlerInvoker(
                 $handlerProvider,
-                $queue,
+                new ResponseWriter($queue),
             ),
         );
     }
