+++ src/FreeDSx/Ldap/Server/Proxy/ProxyProtocolFactory.php
@@ -17,6 +17,7 @@ use FreeDSx\Ldap\LdapClient;
 use FreeDSx\Ldap\Protocol\Authenticator;
 use FreeDSx\Ldap\Protocol\Authorization\DispatchAuthorizer;
 use FreeDSx\Ldap\Protocol\Bind\SimpleBind;
+use FreeDSx\Ldap\Protocol\Queue\Response\ResponseWriter;
 use FreeDSx\Ldap\Protocol\ServerAuthorization;
 use FreeDSx\Ldap\Protocol\ServerProtocolHandler;
 use FreeDSx\Ldap\Protocol\ServerProtocolHandler\ServerStartTlsHandler;
@@ -90,6 +91,7 @@ final class ProxyProtocolFactory implements ServerProtocolFactoryInterface
                     $upstream,
                     $queue,
                 ),
+                new ResponseWriter($queue),
             ),
         );
